Change the algorithm we use to match detections with tracked objects
We used to do our matches so that they would minimize the global
detection-object distance, but found that sometimes this sacrificed
an obvious match with very low distance, for erroneous matches with
higher distance each but which resulted in a lower total distance,
and this created errors in matching.

In order not to leave a detection unmatched the algorithm would
rearrange matches, sacrificing obviously correct matches as this would
allocate matches for more object-detecion pairs and this would minimize
the global distance.

An example of an errenous match caused by this can be seen in the video
"/mnt/hdd1/lalo/videos/in/cu.mp4" @ 1:52 minutes.