Add support for tracking intermittently disappearing points
We have to add the concept of live and not live points for this to work.
Each object has its own life cycle, but internally each point in an
object will now have its own life cycle as well to be able to know which
points from each object are currently being detected and tracked and
which aren't.

In the past we only tracked points which would not dissapear. Which
required us to be careful with filtering detections before feeding them
into the tracker, now this is not a problem anymore. We can just feed
any noisy detection into the tracker and track all the points in it.

Had to modify R to make the velocities on the process model less certain
as it was creating problems for detectors that return (0, 0) for
not detected points. When these points finally get their first real
detection, the tracker sees a huge velocity for the point to have moved
from (0, 0) to wherever they went, and this velocity is then slow to
correct as the following velocity vectors will be much smaller that this
first one. Reducing the velocity certainty in the process model
mitigates this, though I want to futher improve this still in the
future.