Improve distance function in OpenPose demo
We now only compare points which have been detected in the last
detection caught by the tracker. This fixes the problem of points which
haven't been detected for a lot of time and are therefore innacurate to
have incidence in how we calculate our distance function, as it makes it
worse.